refactor(celery): move notification task to job_execution and remove unused notifications module
