+++ backend/app/celery/celery_app.py
@@ -163,8 +163,6 @@ celery_app = Celery(
     backend=settings.REDIS_URL,
     include=[
         "app.celery.tasks.job_execution",
-        "app.celery.tasks.notifications",
-        "app.celery.tasks.evaluation_fast",
     ],
 )
